Move translation mappers to api/Mappers
