+++ api/Functions/RunsUpdateFunction.cs
@@ -9,6 +9,7 @@ using Microsoft.Extensions.Logging;
 using Lfm.Api.Audit;
 using Lfm.Api.Auth;
 using Lfm.Api.Helpers;
+using Lfm.Api.Mappers;
 using Lfm.Api.Middleware;
 using Lfm.Api.Repositories;
 using Lfm.Api.Services;
