+++ tests/Lfm.Api.Tests/GuildMapperTests.cs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 LFM contributors
 
-using Lfm.Api.Functions;
+using Lfm.Api.Mappers;
 using Lfm.Api.Repositories;
 using Lfm.Api.Services;
 using Xunit;
